+++ src/web/components/Settings.tsx
@@ -20,6 +20,7 @@ import {
   TextField,
   Typography,
   Slider,
+  Chip,
 } from "@mui/material";
 import { circularProgress, ScriptApp } from "./Common";
 import {
@@ -87,6 +88,9 @@ export function Settings(params: {
   };
 
   const tickIntervalMsg = `The tool internal update interval is 1 minute, so it may take up to 1 minute ⏳ for some changes to take effect.`;
+  const strengthMin = Math.max(0, cfg.MarketStrengthTargets.min);
+  const strengthMax = Math.min(100, cfg.MarketStrengthTargets.max);
+  const showMarketSlider = false; // TODO: enable on v4.2.0
   return (
     <BasicCard>
       <Stack direction="row" alignItems="center" spacing={1}>
@@ -134,17 +138,21 @@ export function Settings(params: {
           </Stack>
           <FormControl>
             <Typography id="budget-split-slider" gutterBottom>
-              Minimum budget split
+              Budget split
             </Typography>
             <Slider
-              sx={{ ml: 1, width: `96%` }}
+              sx={{ ml: 1, width: `95%` }}
               defaultValue={1}
               value={cfg.BudgetSplitMin}
+              valueLabelDisplay="auto"
               step={1}
-              marks={Array.from({ length: 10 }, (_, i) => ({
-                value: i + 1,
-                label: i + 1,
-              }))}
+              marks={[
+                { value: 1, label: `Auto` },
+                ...Array.from({ length: 9 }, (_, i) => ({
+                  value: i + 2,
+                  label: i + 2,
+                })),
+              ]}
               min={1}
               max={10}
               onChange={(e, newValue) => {
@@ -153,12 +161,77 @@ export function Settings(params: {
               aria-labelledby="budget-split-slider"
             />
             <FormHelperText>
-              This setting limits the budget fraction per trade. A lower value
-              (1) removes the limit and allows to use the entire budget for a
-              single trade, while a higher value (10) allows to allocate only
-              1/10th per trade, reducing the total profit/loss fluctuations.
-              Recommended to increase the limit during unstable markets and
-              reduce during widespread bullish markets.
+              This setting limits the budget fraction per trade. "Auto"
+              dynamically splits the budget into 1-3 parts and can invest the
+              entire budget into a single trade, while a higher value, like 10
+              allows to allocate only 1/10th per trade, allowing to buy more
+              coins. Investing all budget is <b>more risky</b>, but potentially
+              more efficient.
+            </FormHelperText>
+          </FormControl>
+          {showMarketSlider && (
+            <FormControl>
+              <Typography id="market-strength-slider" gutterBottom>
+                <Chip
+                  label="New"
+                  color={`primary`}
+                  size="small"
+                  variant={`outlined`}
+                  style={{ marginRight: 10 }}
+                />
+                Auto-stop (Market strength)
+              </Typography>
+              <Slider
+                sx={{ ml: 1, width: `95%` }}
+                value={[strengthMin, strengthMax]}
+                step={5}
+                min={0}
+                max={100}
+                disableSwap={true}
+                marks={[
+                  { value: 0, label: `0` },
+                  { value: 20, label: `20` },
+                  { value: 40, label: `40` },
+                  { value: 60, label: `60` },
+                  { value: 80, label: `80` },
+                  { value: 100, label: `100` },
+                ]}
+                valueLabelDisplay="auto"
+                onChange={(e, [min, max]: number[]) => {
+                  setCfg({
+                    ...cfg,
+                    MarketStrengthTargets: { min, max },
+                  });
+                }}
+                aria-labelledby="market-strength-slider"
+              />
+              <FormHelperText>
+                This setting controls when to pause trading based on the market
+                {` `}
+                <b>strength</b>. For example, if the range is 10-60, trading
+                starts when the strength reaches 60 or more and stops when it
+                drops below 10. It resumes only after the strength reaches 60
+                again.
+              </FormHelperText>
+            </FormControl>
+          )}
+          <FormControl>
+            <FormControlLabel
+              control={
+                <Switch
+                  checked={cfg.SmartExit}
+                  onChange={(e) => {
+                    setCfg({ ...cfg, SmartExit: e.target.checked });
+                  }}
+                />
+              }
+              label="Smart exit"
+              aria-describedby={`smart-exit-helper-text`}
+            />
+            <FormHelperText id={`smart-exit-helper-text`}>
+              Sell automatically when conditions are no longer in favor and it
+              is better to take the profit/loss. Recommended to always keep
+              enabled. Disable only if you need to hold the assets.
             </FormHelperText>
           </FormControl>
           <FormControl>
@@ -185,25 +258,6 @@ export function Settings(params: {
               >
                 https://binance.com/en/fee
               </Link>
-            </FormHelperText>
-          </FormControl>
-          <FormControl>
-            <FormControlLabel
-              control={
-                <Switch
-                  checked={cfg.SmartExit}
-                  onChange={(e) => {
-                    setCfg({ ...cfg, SmartExit: e.target.checked });
-                  }}
-                />
-              }
-              label="Smart exit"
-              aria-describedby={`smart-exit-helper-text`}
-            />
-            <FormHelperText id={`smart-exit-helper-text`}>
-              Sell automatically when conditions are no longer in favor and it
-              is better to take the profit/loss. Recommended to always keep
-              enabled. Disable only if you need to hold the assets.
             </FormHelperText>
           </FormControl>
           <FormControl>